requirements.txt
sqlalchemy-spanner==1.7.0
